Working configuration with Linux 4.19 and Python3
This image boots when the .hddimage is dd'd onto a USB stick and booted. There is an option to install this configuration to the internal memory. Python3 is installed and can be started. The iwd service does not start since the kernel version is less than 4.2. Wifi driver appears to work and can scan nearby stations.